leak fixed

+++ atoi.c
@@ -6,13 +6,13 @@
 /*   By: satabay <satabay@student.42istanbul.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/12 13:39:51 by satabay           #+#    #+#             */
-/*   Updated: 2025/10/18 13:35:59 by satabay          ###   ########.fr       */
+/*   Updated: 2025/10/18 17:32:35 by satabay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(const char *nptr)
+long long int	ft_atoi(const char *nptr)
 {
 	int				sign;
 	long long int	num;
